@@ -7,15 +7,14 @@
   "license": "N/A",
   "private": true,
   "scripts": {
-    "dev": "concurrently \"tsc --pretty --watch\" \"next --inspect\"",
-    "prebuild": "tsc",
+    "dev": "next --inspect",
     "build": "next build",
     "start": "next start",
-    "analyze": "cross-env ANALYZE=1 next build"
+    "analyze": "cross-env BUNDLE_ANALYZE=both next build"
   },
   "dependencies": {
     "moment": "^2.20.1",
-    "next": "^4.4.0-canary.3",
+    "next": "^5.0.0",
     "prop-types": "^15.6.0",
     "react": "^16.2.0",
     "react-bootstrap": "^0.32.1",
@@ -24,26 +23,18 @@
   },
   "devDependencies": {
     "@types/moment": "^2.13.0",
-    "@types/next": "^2.4.7",
+    "@types/next": "^2.4.8",
     "@types/prop-types": "^15.5.2",
-    "@types/react": "^16.0.34",
+    "@types/react": "16.0.34",
     "@types/react-bootstrap": "^0.32.1",
     "@types/react-ga": "^2.3.0",
     "@types/styled-jsx": "^2.2.1",
-    "autoprefixer": "^7.2.5",
-    "babel-plugin-module-resolver": "^3.0.0",
-    "babel-plugin-wrap-in-js": "^1.1.1",
-    "concurrently": "^3.5.1",
+    "@zeit/next-bundle-analyzer": "^0.0.4",
+    "@zeit/next-sass": "^0.1.1",
+    "@zeit/next-typescript": "^0.0.10",
     "cross-env": "^5.1.3",
-    "glob": "^7.1.2",
     "node-sass": "^4.7.2",
-    "normalize.css": "^7.0.0",
-    "postcss-easy-import": "^3.0.0",
-    "postcss-loader": "^2.0.10",
-    "raw-loader": "^0.5.1",
-    "sass-loader": "^6.0.6",
     "tslint": "^5.9.1",
-    "typescript": "^2.6.2",
-    "webpack-bundle-analyzer": "^2.9.2"
+    "typescript": "^2.7.2"
   }
 }
